+++ Demo/Keyboard/DemoActionHandler.swift
@@ -102,10 +102,7 @@ private extension UIImage {
 }
 
 
-/**
- This class is used as a target/selector holder by the image
- extension above.
- */
+/// This class is used as target by the extension above.
 private class ImageService: NSObject {
     
     public typealias Completion = (Error?) -> Void