+++ RELEASE_NOTES.md
@@ -23,10 +23,18 @@ Another autocomplete update is that the `KeyboardInputViewController` will check
 
 Another change worth knowing, is that the `Autocomplete.ToolbarItem` view no longer adds quotations around unknown suggestions. That responsibility has been moved to the autocomplete provider. The view will just show the titles it receives.  
 
+### 🆕 Settings
+ 
+* `AutocompleteSettings` is a new observable settings type.
+* `DictationSettings` is a new observable settings type.
+* `FeedbackSettings` is a new observable settings type.
+* `Keyboard.Settings` is a new settings container type.
+
 ### ✨ Features
 
 * `Autocomplete.Suggestion` has new functions.
 * `AutocompleteContext` has a new `isAutoLearnEnabled` property.
+* `AutocompleteContext` has a new `preferredSuggestionCount` property.
 * `AutocompleteProvider` has new `ignoreWords(_:)` and suggestion functions.
 * `KeyboardAction.StandardProvider` can now automatically learn unknown suggestions.
 * `KeyboardInputViewController` has a new `settings` property with setting instances.